Bugs in MLHandler fixed. Some bugs in DIHandler have been fixed to.
One array had the wrong size shich caused the program to write out of bounds
and therefore cause a heap corruption. The parsing of the keypresses in
MLHandler were done in the same way as in the DI. AS they represent different
ways of handling input this malfunctioned. A new calcStateFromEvents has been
written which calculates the correct state based on events fired (as in ML)
instead of the current key state (as in DI). Mouse position is still done
using ML all the time.